requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Your existing dependencies (from your current system)
openai==1.3.8
sentence-transformers==2.2.2
pinecone-client==2.2.4
numpy==2.2.6
opencv-python==4.12.0.90

# Video processing
moviepy==1.0.3
webvtt-py==0.4.6

# Data handling
pandas==2.1.4
pydantic==2.5.1
aiofiles==23.2.1
tqdm==4.66.1

# Additional dependencies that might be needed
python-dateutil==2.8.2

# Development tools
pytest==7.4.3
httpx==0.25.2

# Additional dependencies you might already have installed
scikit-learn
matplotlib
seaborn